Rename
Rename test_Lambda.{c,h} to test_lnLambda.{c,h}.
