--- src/tests/test_Lambda.h
@@ -1,6 +1,0 @@
-#ifndef TEST_LAMBDA_H
-#define TEST_LAMBDA_H
-
-int test_caps_lnLambda(void);
-
-#endif
