+++ src/tests/test_lnLambda.h
@@ -1,0 +1,6 @@
+#ifndef TEST_LNLAMBDA_H
+#define TEST_LNLAMBDA_H
+
+int test_caps_lnLambda(void);
+
+#endif
